Aight All features done. I can now waste my life working on UI

+++ app/src/main/java/com/r19012817/weatherapp/CityQueryPackage/TimeZone.java
@@ -5,7 +5,7 @@ public class TimeZone {
 
     private String Name;
 
-    private int GmtOffset;
+    private double GmtOffset;
 
     private boolean IsDaylightSaving;
 
@@ -31,7 +31,7 @@ public class TimeZone {
         this.GmtOffset = GmtOffset;
     }
 
-    public int getGmtOffset() {
+    public double getGmtOffset() {
         return this.GmtOffset;
     }
